Extract repo copy box
